Updated functional control icons to 24x24.
Refactored Anchor, Button, and Menu accordingly.

This change was necessary to allow the icons to be used more broadly.
We removed the 12px "padding" rect within each icon. The visible part
of the icons did not change, just the space around them. Now, the containing
Button and Anchor components take care of adding needed padding around
any icons used with them.

A notable change in Menu is that the control when the Menu is collapsed
now uses a Button.

And, Button now supports having an icon and label.

NOTE: These changes might impact how your Button and Anchor components look.
We have done a variety of testing so that these changes are non-impacting
but we recommend evaluating your usage just to be sure.