Fix: add a11yTitle prop to social share
The a11yTitle that was automatically set was failing chrome accessibility audit
due to AX_TEXT_04.  To enable overriding by the end user, the a11yTitle can be
passed in as an optional prop.  If not supplied, then a default a11yTitle is set
as Share on Type where Type is the social network.

SEE: https://github.com/GoogleChrome/accessibility-developer-tools/wiki/Audit-Rules#ax_text_04