Allow columns to be individually sortable.
Column labels now allow for an optional tuple to define a column header.
The first item is just the column label (React node). The second item
allows for an options object. This only supports the `sortable` option
to disable sorting for some columns. If no options are passed the
default is still to rely on the `onSort` value to enable the sorting UI.