Added defaultMessage as an argument
Adding "defaultMessage" argument to the getMessage function so that the defaultMessage value is not always same as the key argument value which is passed to the getMessage function. 
Due to the absence of "defaultMessage" argument,  developers are forced to keep the value of key argument as what they want the defaultMessage to be shown. Devs want to name their keys in a unique manner and sometimes component specific however the current code restricts it as only key argument is present in the function.