Issue 2160 - Enhancement to specify the DOM element to create the Menu drop down in (#2161)
If a developer wants to place a Menu within a floating panel (tooltip,etc) its very helpful to constrain the Menu's floating drop down to be rendered in the DOM of the floating parent. This makes life cycle much easier (Menu won't hang around after floating parent is dismissed) and it makes it easier to cause the menu to render within the parent.