fix(Paragraph): Removes console warning for passing `fill` to `p`
In #3222, we added the `fill` property. Styled Components passes all
properties through to their DOM children. However, because `fill` isn't
in react's `camelCase` format, React assumes this must be a native
DOM property and throws a warning ("Received `true` for a non-boolean
attribute `fill`.)

The fix means consuming the `fill` prop on `Paragraph` and converting
it to `fillProp` when calling the `StyledParagraph` object. This is
similar to the fix we use for `color` / `colorProp`.

Signed-off-by: Jakob Heuser <jakob+npm@felocity.com>