Update README.md / Added standalone componnets
Added a link to bit.dev/grommet/grommet for the installation of standalone components, in the micro-frontends/atomic-design spirit.
It also lets people import the components into their own projects for modifications when needed. Even though there are examples there too, I added it under "Installation" and not "Explore" since the real value is in the flexible distribution and consumption of standalone components.