Fix Button propTypes for the target property to allow string (#4236)
The Button propTypes currently complain about specifying
something other than _self, _blank, _parent, _top. In
html the anchor tag actually allows any string in target
to create a named tab/window. The names above that start
with _ are special names but it is perfectly fine to
create a named window.

This is a fix for issue #4097 in grommet.