Support clicking inside drops in a shadow root
If a drop is rendered inside a shadow root, clicks in the drop were
still triggering onClickOutside. This is is because the host element
for the shadow root is the target on the document mousedown event
making it appear to be outside of the drop.

When a shadow root is attached in open mode, it is possible to get
the real target by using event.composedPath(). This change uses
composedPath to get the target if it is supported, and gracefully
falls back to event.target when it isn't supported.

Signed-off-by: Brent Burgoyne <brent47@gmail.com>