Fix MaskedInput 'fixed' segments which are longer than 1 character
Anytime a mask segment is a 'fixed' string of more than 1
character, it always adds the full fixed segment in an input value.
However when the user attempts to backspace over that fixed
value, once they back over the last character in that segment,
MaskedInput puts back in the entire fixed string twice (minus the last
character of the second copy.)

The change allows you to have just part of the fixed segment there
until you add something that doesn't match the the first part of the
fixed string at which point it fills in the rest of the fixed part for
you. This works much more naturally with general typing and backspace.
It also still works the same as it used to for the more typical
single character fixed portions (like the phone number or IP Address
examples).

I added a storybook example of a multi-character fixed string (URL)
since it seemed like a useful example either way.

Fixes #4122