InfiniteScroll stops intermittently
The pageHeight is calculated at the beginning of the rendered component but pageHeight is kept constant. A page can have varying content as a user scrolls and pageHeight is still kept the same and can cause the InfiniteScroll to miscalculate the nextEndPage. Rather than using pageHeight, I am proposing the usage of scrollHeight to determine when to call our onMore() function.

Signed-off-by: mrnguuyen <nnbrandon@att.net>