borders: Support passing arrays to allow per side border config.
BorderType suggests that it supports arrays to allow configuring each
side of the border independently of the other sides, however, when
actually trying to use an array of border objects, we find that they are
not respected. This commit adds support for arrays by recursively
merging each border entry and returns that. Added a storybook entry to
demonstrate this.