Added a "justify" to the ExpanderControl component to make the icon vertically centered and to handle non-standard themes or styling.
Removed the "usingKeyboard", it conflicted with switching apps (at least on mac) using Cmd-Tab and rendered the top row focused/active independent on which rows was expanded. The only case I could find where this would be applicable would be using the "onSelect" use-case, but that still works as intended after manual testing. I would appreciate some input whether which is correct though.