Fixes #7418 Add onClickOutside handler for Tip component
Making clone of children in slow processing environment
leads to a situation where mouseEnter event happens for
a previous element and so for a new cloned element that
is trigger for Tip component, mouseLeave never triggers
leaving Tip component hanging. Although the situation may
not be improved however a missing prop of onClickOutside
is not used yet, so by adding this prop we eliminate this
problem if user clicks anywhere on document.

Signed-off-by: Priya Ranjan Singh <priya-ranjan.singh@hpe.com>